Password reset page ui done
made the ui for the password reset page, button clicks through to it on login page.

also added the actual password reset page ui for when people click on the password reset in their emails, ready for later. the ui for this works and you can test that using this url

http://localhost:3001/reset-password/test-token-123